+++ src/extension.ts
@@ -11,6 +11,7 @@ import { registerHoiFs } from './util/hoifs';
 import { loadI18n } from './util/i18n';
 import { registerGfxIndex } from './util/gfxindex';
 import { registerLocalisationIndex } from "./util/localisationIndex";
+import { registerLocalisationHighlighting } from './util/localisationHighlighting';
 import { registerSharedFocusIndex } from "./util/sharedFocusIndex";
 
 export function activate(context: vscode.ExtensionContext) {
@@ -37,6 +38,7 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(registerSharedFocusIndex());
     context.subscriptions.push(registerGfxIndex());
     context.subscriptions.push(registerLocalisationIndex());
+    context.subscriptions.push(registerLocalisationHighlighting());
 
     if (process.env.NODE_ENV !== 'production') {
         vscode.commands.registerCommand('hoi4modutilities.test', () => {